# arinar-v2/packages/contracts/openapi/arinar-v1.yaml
openapi: 3.1.0
info:
  title: Arinar API
  version: 1.0.0
  description: |
    Arinar V2 API - AI-native knowledge platform with multi-agent realtime discussions.
    
    **Model Provider Policy**: This system uses OpenRouter for LLM access, not direct OpenAI/Anthropic integrations.
  contact:
    name: Arinar Platform Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.arinar.dev
    description: Development environment
  - url: https://api.arinar.com
    description: Production
...
paths:
...
  /debates:
    post:
      operationId: createDebate
      summary: Create new debate session
      description: Creates a new debate session in pending state (M2)
      tags:
        - debates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDebateRequest'
      responses:
        '201':
          description: Debate created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebateResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debates/{debate_id}:
    get:
      operationId: getDebate
      summary: Get debate details
      description: Retrieve details for a specific debate (M2)
      tags:
        - debates
      parameters:
        - $ref: '#/components/parameters/DebateIdParam'
      responses:
        '200':
          description: Debate retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebateResponse'
        '404':
          description: Debate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debates/run:
    post:
      operationId: runDebate
      summary: Run debate-in-a-box (M1)
      description: |
        Executes a complete 5-turn round-robin debate with 3 agents.
        Returns summary, minutes of meeting, action items, and full event history.
        
        **OpenRouter BYOK**: Requires user-provided OpenRouter API key per request.
      tags:
        - debates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebateRunRequest'
      responses:
        '200':
          description: Debate completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebateRunResponse'
        '400':
          description: Invalid request (wrong number of agents, missing fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid OpenRouter API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debates/setup:
    post:
      operationId: setupDebate
      summary: Setup a debate (participants + materials)
      description: Creates a pending debate plus participants and materials metadata in one call.
      tags:
        - setup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebateSetupRequest'
      responses:
        '201':
          description: Debate setup created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebateSetupResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /debates/{debate_id}/start:
    post:
      operationId: startDebate
      summary: Start debate session
      description: Transitions debate from draft/preflight to live state
      tags:
        - debates
      parameters:
        - $ref: '#/components/parameters/DebateIdParam'
      responses:
        '200':
          description: Debate started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebateResponse'
        '404':
          description: Debate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debates/{debate_id}/pause:
    post:
      operationId: pauseDebate
      summary: Pause debate (M2)
      description: Transitions debate from running to paused state
      tags:
        - debates
      parameters:
        - $ref: '#/components/parameters/DebateIdParam'
      responses:
        '200':
          description: Debate paused successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebateResponse'
        '400':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Debate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debates/{debate_id}/resume:
    post:
      operationId: resumeDebate
      summary: Resume debate (M2)
      description: Transitions debate from paused to running state
      tags:
        - debates
      parameters:
        - $ref: '#/components/parameters/DebateIdParam'
      responses:
        '200':
          description: Debate resumed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebateResponse'
        '400':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Debate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debates/{debate_id}/intervene:
    post:
      operationId: interveneDebate
      summary: Intervene in debate
      description: Manual or policy-triggered intervention in active debate
      tags:
        - debates
      parameters:
        - $ref: '#/components/parameters/DebateIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterventionRequest'
      responses:
        '200':
          description: Intervention created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionResponse'
        '404':
          description: Debate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debates/{debate_id}/end:
    post:
      operationId: endDebate
      summary: End debate session
      description: Transitions debate to synthesis/closed state
      tags:
        - debates
      parameters:
        - $ref: '#/components/parameters/DebateIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndDebateRequest'
      responses:
        '200':
          description: Debate ended successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebateResponse'
        '404':
          description: Debate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debates/{debate_id}/summarize:
    post:
      operationId: generateSummary
      summary: Generate debate summary (M3)
      description: |
        Generate end-of-meeting outputs: summary, minutes, action items.
        Requires debate to be in 'ended' state.
        
        **OpenRouter BYOK**: Requires user-provided OpenRouter API key per request (never stored).
      tags:
        - debates
      parameters:
        - $ref: '#/components/parameters/DebateIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SummarizeRequest'
      responses:
        '200':
          description: Summary generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummaryResponse'
        '400':
          description: Invalid request (debate not ended)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid OpenRouter API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Debate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debates/{debate_id}/summary:
    get:
      operationId: getSummary
      summary: Get debate summary (M3)
      description: Retrieve existing summary/minutes/action items for a debate
      tags:
        - debates
      parameters:
        - $ref: '#/components/parameters/DebateIdParam'
      responses:
        '200':
          description: Summary retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummaryResponse'
        '404':
          description: Debate or summary not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    DebateIdParam:
      name: debate_id
      in: path
      description: Debate session UUID
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    CreateDebateRequest:
      type: object
      required:
        - workspace_id
        - title
      properties:
        workspace_id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 500
        policy_config:
          type: object
          additionalProperties: true
    DebateResponse:
      type: object
      required:
        - debate_id
        - workspace_id
        - title
        - state
        - created_at
      properties:
        debate_id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        title:
          type: string
        state:
          $ref: '#/components/schemas/DebateState'
        created_at:
          type: string
          format: date-time
    DebateState:
      type: string
      enum:
        - pending
        - running
        - paused
        - ended
...
    DebateSetupRequest:
      type: object
      required:
        - workspace_id
        - title
        - problem_statement
        - participants
      properties:
        workspace_id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 500
        problem_statement:
          type: string
          minLength: 1
        timebox_minutes:
          type: integer
          minimum: 1
        participants:
          type: array
          minItems: 1
          maxItems: 8
          items:
            $ref: '#/components/schemas/SetupParticipant'
        materials:
          type: array
          items:
            $ref: '#/components/schemas/SetupMaterial'
...
    DebateSetupResponse:
      type: object
      required:
        - debate_id
        - participant_ids
        - material_ids
      properties:
        debate_id:
          type: string
          format: uuid
        participant_ids:
          type: array
          items:
            type: string
            format: uuid
        material_ids:
          type: array
          items:
            type: string
            format: uuid
    DebateRunRequest:
      type: object
      required:
        - problem_statement
        - agents
        - openrouter_api_key
      properties:
        problem_statement:
          type: string
          description: Problem or topic to discuss
          minLength: 10
          maxLength: 5000
        agents:
          type: array
          description: Exactly 3 agents for M1
          minItems: 3
          maxItems: 3
          items:
            $ref: '#/components/schemas/AgentInput'
        openrouter_api_key:
          type: string
          description: OpenRouter API key (BYOK - Bring Your Own Key)
          minLength: 1
        debate_title:
          type: string
          description: Optional debate title
          default: "Untitled Debate"
          maxLength: 500
...
    DebateRunResponse:
      type: object
      required:
        - debate_id
        - status
        - outputs
        - event_history
      properties:
        debate_id:
          type: string
          format: uuid
          description: Generated debate session ID
        status:
          type: string
          enum: [completed, failed]
          description: Final debate status
        outputs:
          type: object
          required:
            - summary
            - minutes_of_meeting
            - action_items
          properties:
            summary:
              type: string
              description: High-level summary (2-3 sentences)
            minutes_of_meeting:
              type: string
              description: Detailed minutes of meeting (200-300 words)
            action_items:
              type: array
              description: Concrete action items
              items:
                type: string
        event_history:
          type: array
          description: Ordered list of debate events (turns)
          items:
            type: object
            required:
              - event_id
              - turn
              - agent
              - message
            properties:
              event_id:
                type: string
                format: uuid
              turn:
                type: integer
                minimum: 1
                maximum: 5
              agent:
                type: string
              message:
                type: string
...
    InterventionRequest:
      type: object
      required:
        - intent
        - question_or_instruction
      properties:
        intent:
          type: string
          enum:
            - clarify
            - challenge
            - re_scope
            - request_evidence
            - decision_check
        target_agents:
          type: array
          items:
            type: string
            format: uuid
        question_or_instruction:
          type: string
          minLength: 1
          maxLength: 2000
        blocking:
          type: boolean
          default: false
          description: If true, pauses normal queue until addressed
    InterventionResponse:
      type: object
      required:
        - intervention_id
        - debate_id
        - created_at
      properties:
        intervention_id:
          type: string
          format: uuid
        debate_id:
          type: string
          format: uuid
        intent:
          type: string
        target_agents:
          type: array
          items:
            type: string
            format: uuid
        question_or_instruction:
          type: string
        blocking:
          type: boolean
        created_at:
          type: string
          format: date-time
    SummarizeRequest:
      type: object
      required:
        - openrouter_api_key
      properties:
        openrouter_api_key:
          type: string
          minLength: 10
          description: OpenRouter BYOK key (never stored)
        model_id:
          type: string
          default: "anthropic/claude-3.5-sonnet"
          description: Model for summary generation
...
    EndDebateRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
...